% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLDA.R
\name{CTLDA}
\alias{CTLDA}
\title{CTLDA}
\usage{
CTLDA(pvec)
}
\arguments{
\item{pvec}{a vector a p-values to be combined}
}
\value{
a numerical value of p-value from Cauchy Combination Test
}
\description{
CTLDA: A Combined Test for Longitudinal Data Analysis in Randomized Clinical Trials using Cauchy Combination Test
}
\references{
Zhai, S., Shen, J. and Mehrotra, D.V. (2016). CTLDA: A Combined Test for Longitudinal Data Analysis in Randomized Clinical Trials.
}
\author{
Song Zhai
}
