% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prop_delay.R
\name{f.prop.delay.m12}
\alias{f.prop.delay.m12}
\title{f.prop.delay.m12}
\usage{
f.prop.delay.m12(t, v0, v1, v2, v3, v4, b)
}
\arguments{
\item{t}{time}

\item{v0}{mean response at visit 0 in the placebo arm}

\item{v1}{mean response at visit 1 in the placebo arm}

\item{v2}{mean response at visit 2 in the placebo arm}

\item{v3}{mean response at visit 3 in the placebo arm}

\item{v4}{mean response at visit 4 in the placebo arm}

\item{b}{parameter of interest}
}
\value{
mean response in the active arm
}
\description{
PMRM (proportional delay) core function
}
\references{
Raket, L.L. (2022). Progression models for repeated measures: Estimating novel treatment effects in progressive diseases. Statistics in Medicine, 41(28), 5537-5557.
}
\author{
Song Zhai
}
