% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRM.R
\name{MMRM}
\alias{MMRM}
\title{MMRM}
\usage{
MMRM(dat, duration)
}
\arguments{
\item{dat}{a data frame of standard longitudinal data}

\item{duration}{a numerical value of study duration}
}
\value{
a numeric list including fitted linear mixed effects model and contrast test results
}
\description{
MMRM (last-visit)
}
\references{
Detry, M.A. and Ma, Y. (2016). Analyzing repeated measurements using mixed models. Jama, 315(4), 407-408.
}
\author{
Song Zhai
}
