% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCS.R
\name{ncs_ranslp_df2}
\alias{ncs_ranslp_df2}
\title{ncs_ranslp_df2}
\usage{
ncs_ranslp_df2(dat, intercept = TRUE)
}
\arguments{
\item{dat}{a data frame of standard longitudinal data}

\item{intercept}{an indicator whether or not to include intercept}
}
\value{
a numeric list including fitted natural cubic spline model and Wald test results
}
\description{
Natural Cubic Spline
}
\references{
Donohue, M.C., Langford, O., Insel, P.S., et al. (2023). Natural cubic splines for the analysis of Alzheimer's clinical trials. Pharmaceutical statistics, 22(3), 508-519
}
\author{
Song Zhai
}
