% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pAUC.R
\name{pAUC}
\alias{pAUC}
\title{pAUC}
\usage{
pAUC(dat, start, nboot = 1000)
}
\arguments{
\item{dat}{a data frame of standard longitudinal data}

\item{start}{a numerical value of starting visit point to calculate the area}

\item{nboot}{number of bootstraps}
}
\value{
a list of the observed area, null distribution of the area, and the p-value
}
\description{
pAUC (partial AUC): pAUC measures the treatment effect as the area between active and placebo time-to-response curves over the post-baseline visits
}
\references{
Atri, A., Hendrix, S.B., Pejovic, V., et al. (2015). Cumulative, additive benefits of memantine-donepezil combination over component monotherapies in moderate to severe Alzheimer’s dementia: a pooled area under the curve analysis. Alzheimer's Research & Therapy, 7, 1-12.
}
\author{
Song Zhai
}
