% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMRM_prop_delay.R
\name{prop.delay.m12}
\alias{prop.delay.m12}
\title{prop.delay.m12}
\usage{
prop.delay.m12(dat)
}
\arguments{
\item{dat}{a data frame of standard longitudinal data}
}
\value{
a list of fitted progression model for repeated measures and p-value
}
\description{
PMRM (proportional delay)
}
\references{
Raket, L.L. (2022). Progression models for repeated measures: Estimating novel treatment effects in progressive diseases. Statistics in Medicine, 41(28), 5537-5557.
}
\author{
Song Zhai
}
