% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_corr.R
\name{simulate_corr}
\alias{simulate_corr}
\title{simulate_corr}
\usage{
simulate_corr(pattern, M)
}
\arguments{
\item{pattern}{a character of desired variance-covariance pattern name}

\item{M}{a vector of visit points}
}
\value{
a list of variance-covariance matrix and correlation matrix
}
\description{
Simulate variance-covariance matrix and correlation matrix
}
\references{
Zhai, S., Shen, J. and Mehrotra, D.V. (2016). CTLDA: A Combined Test for Longitudinal Data Analysis in Randomized Clinical Trials.
}
\author{
Song Zhai
}
