% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_TE.R
\name{simulate_TE}
\alias{simulate_TE}
\title{simulate_TE}
\usage{
simulate_TE(f, M, pattern)
}
\arguments{
\item{f}{time-to-response function in the placebo arm}

\item{M}{a numerical vector of visit points}

\item{pattern}{a character of desired treatment effect pattern}
}
\value{
a numerical vector of mean responses at visit points in the active arm
}
\description{
Simulate treatment effect pattern
}
\references{
Zhai, S., Shen, J. and Mehrotra, D.V. (2016). CTLDA: A Combined Test for Longitudinal Data Analysis in Randomized Clinical Trials.
}
\author{
Song Zhai
}
