% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{spline_nonpara}
\alias{spline_nonpara}
\title{spline_nonpara}
\usage{
spline_nonpara(ys.train, xs.train, xs.test)
}
\arguments{
\item{ys.train}{mean responses in the training set}

\item{xs.train}{visit points in the trauning set}

\item{xs.test}{visit points in the testing set}
}
\value{
mean responses in the testing set
}
\description{
Spline function
}
\references{
Raket, L.L. (2022). Progression models for repeated measures: Estimating novel treatment effects in progressive diseases. Statistics in Medicine, 41(28), 5537-5557.
}
\author{
Song Zhai
}
